+++ src/modules/resources/extraction/index.ts
@@ -2,3 +2,4 @@ export * from './types';
 export * from './extractors';
 export { BrowserExtractionService, browserExtractionService } from './BrowserExtractionService';
 export { AIExtractionService, aiExtractionService } from './AIExtractionService';
+export { DocumentTreeExtractor } from './documentTree';